grsec: Apply grsecurity-2.2.1-2.6.36.1-201012091553.patch
Signature-tree: 1017635e7d7f34b8faa14b0c9e8c3eb9aa6ca94d

+++ drivers/pci/probe.c
@@ -165,7 +165,7 @@ int __pci_read_base(struct pci_dev *dev, enum pci_bar_type type,
 	u32 l, sz, mask;
 	u16 orig_cmd;
 
-	mask = type ? PCI_ROM_ADDRESS_MASK : ~0;
+	mask = type ? (u32)PCI_ROM_ADDRESS_MASK : ~0;
 
 	if (!dev->mmio_always_on) {
 		pci_read_config_word(dev, PCI_COMMAND, &orig_cmd);
